+++ glibc/pthread/samples/sys_affinity_bind.h
@@ -1,2 +1,0 @@
-#pragma once
-int sys_affinity_bind(int cpu);